+++ tests/query.test.ts
@@ -123,7 +123,9 @@ class Count extends System {
 }
 
 async function createWorld(...systems: SystemType<System>[]): Promise<World> {
-  return World.create({maxEntities: 100, defs: [systems, Count]});
+  return World.create({
+    maxEntities: 100, defaultComponentStorage: 'sparse', defs: [systems, Count]
+  });
 }
 
 
@@ -229,7 +231,8 @@ describe('creating and deleting entities', () => {
 
   test('recycle entity IDs', async() => {
     const world = await World.create({
-      maxEntities: 9, maxLimboEntities: 7, defs: [CreateAForEachC, DeleteA]
+      maxEntities: 9, maxLimboEntities: 7, defaultComponentStorage: 'sparse',
+      defs: [CreateAForEachC, DeleteA]
     });
     world.createEntity(C, {value: 1});
     world.createEntity(C, {value: 2});